refactor: resolve circular dependencies and extract service registration
- Created `ServicesManager` in `custom_components/meraki_ha/services/manager.py` to encapsulate service registration logic.
- Updated `__init__.py` to instantiate `ServicesManager` and register services.
- Removed `async_setup_services` from `MerakiDataUpdateCoordinator` in `coordinator.py` and removed service imports.
- Deleted `custom_components/meraki_ha/core/coordinator_helpers/service_setup.py`.
- Verified with scorecard and tests.
